Remove branch
