+++ package.json
@@ -3,6 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "babel-eslint": "^8.2.3",
     "flexboxgrid2": "^7.2.1",
     "normalize.css": "^8.0.0",
     "prettier": "1.13.5",
